+++ lib/odd_job/async.ex
@@ -2,6 +2,7 @@ defmodule OddJob.Async do
   @moduledoc false
   @moduledoc since: "0.1.0"
   alias OddJob.{Job, Utils}
+  import OddJob.Utils, only: [pool_name: 1, proxy_sup_name: 1]
 
   @type job :: OddJob.Job.t()
 
@@ -10,7 +11,7 @@ defmodule OddJob.Async do
   @spec perform(atom, fun) :: job
   def perform(pool, fun) when is_atom(pool) and is_function(fun) do
     pool
-    |> supervisor()
+    |> proxy_sup_name()
     |> DynamicSupervisor.start_child(@server)
     |> Utils.extract_pid()
     |> Utils.link_and_monitor()
@@ -23,7 +24,7 @@ defmodule OddJob.Async do
     jobs =
       for item <- collection do
         pool
-        |> supervisor()
+        |> proxy_sup_name()
         |> DynamicSupervisor.start_child(@server)
         |> Utils.extract_pid()
         |> Utils.link_and_monitor()
@@ -31,13 +32,13 @@ defmodule OddJob.Async do
         |> send_job_to_proxy()
       end
 
-    GenServer.cast(pool, {:perform_many, jobs})
+    GenServer.cast(pool_name(pool), {:perform_many, jobs})
     jobs
   end
 
   @spec run_proxy_with_job(job, atom) :: job
   defp run_proxy_with_job(job, pool) do
-    GenServer.call(job.proxy, {:run, pool, job})
+    GenServer.call(job.proxy, {:run, pool_name(pool), job})
   end
 
   @spec build_job({pid, reference}, function) :: job
@@ -56,8 +57,6 @@ defmodule OddJob.Async do
     GenServer.cast(job.proxy, {:job, job})
     job
   end
-
-  defp supervisor(pool_id), do: OddJob.Supervisor.proxy_sup_name(pool_id)
 
   # The rest of this module, covering implementation of the `await` and `await_many` functions,
   # was copied and adapted from the analogous functions in the Elixir standard library's `Task` module.